backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
supabase==2.3.4
sqlalchemy==2.0.25
asyncpg==0.29.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Background Tasks
celery[redis]==5.3.6

# Payments
requests==2.31.0

# File Processing
Pillow==10.2.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
potrace==1.8

# Security
python-dotenv==1.0.0
httpx==0.26.0

# Validation
email-validator==2.1.0.post1
pydantic==2.5.3
pydantic-settings==2.1.0
